requirements.txt
numpy==1.19.5
h2o==3.30.0.4
pandas==1.3.0
scikit-learn==0.23.2
matplotlib==3.4.2
google-cloud-bigquery-storage==2.6.0
google-cloud-bigquery==2.21.0
pyarrow==4.0.1